requirements.txt
crewai
crewai-tools
openai
requests
python-dotenv
pyyaml